+++ pf-play-rule/src/main/java/com/pf/play/rule/core/model/UMasonryListLog.java
@@ -50,7 +50,7 @@ public class UMasonryListLog {
      *
      * @mbggenerated
      */
-    private Float masonryNum;
+    private Double masonryNum;
 
     /**
      * 赠送人
@@ -153,11 +153,11 @@ public class UMasonryListLog {
         this.symbolType = symbolType;
     }
 
-    public Float getMasonryNum() {
+    public Double getMasonryNum() {
         return masonryNum;
     }
 
-    public void setMasonryNum(Float masonryNum) {
+    public void setMasonryNum(Double masonryNum) {
         this.masonryNum = masonryNum;
     }
